+++ src/platform/react-native/index.ts
@@ -8,7 +8,7 @@ import BufferUtils from '../web/lib/util/bufferutils';
 // @ts-ignore
 import Crypto from '../web/lib/util/crypto';
 import Http from '../web/lib/util/http';
-import Config from './platform';
+import Config from './config';
 // @ts-ignore
 import Transports from '../web/lib/transport';
 import Logger from '../../common/lib/util/logger';
