+++ src/platform/web/config-webworker.ts
@@ -1,0 +1,5 @@
+import Config from './config';
+
+Config.isWebworker = true;
+
+export default Config;
