+++ src/platform/web/index-webworker.ts
@@ -1,4 +1,4 @@
-import './platform-webworker';
+import './config-webworker';
 import Ably from './index';
 
 export default Ably;
