--- src/platform/web/platform-webworker.ts
@@ -1,5 +1,0 @@
-import Platform from './platform';
-
-Platform.isWebworker = true;
-
-export default Platform;
